move get_last_two_message_ids to feedback_services
