+++ core/domain/feedback_domain.py
@@ -119,19 +119,6 @@ class FeedbackThread(python_utils.OBJECT):
         """
         return '.'.join([self.id, python_utils.UNICODE(message_id)])
 
-    def get_last_two_message_ids(self):
-        """Returns the full message ids of the last two messages of the thread.
-        If the thread has only one message, the id of the second last message is
-        None.
-
-        Returns:
-            list(str|None). The ids of the last two messages of the thread. If
-                the message does not exist, None is returned.
-        """
-        return [self.get_full_message_id(i) if i >= 0 else None
-                for i in range(
-                    self.message_count - 1, self.message_count - 3, -1)]
-
 
 class FeedbackMessage(python_utils.OBJECT):
     """Domain object for a feedback message.
